pyproject.toml
[project]
name = "rent-notification-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]
